Implement JSON-ld signature verification
Implement JSON-ld signature verification
Move signature verification functions to the proper file
