+++ plume-common/src/activity_pub/mod.rs
@@ -1,10 +1,8 @@
 use activitypub::{Activity, Actor, Object, Link};
 use array_tool::vec::Uniq;
-use base64;
 use reqwest::Client;
 use rocket::{
-    Outcome,
-    http::{Status,HeaderMap},
+    Outcome, http::Status,
     response::{Response, Responder},
     request::{FromRequest, Request}
 };
@@ -117,68 +115,6 @@ pub fn broadcast<S: sign::Signer, A: Activity, T: inbox::WithInbox + Actor>(send
             Ok(mut r) => println!("Successfully sent activity to inbox ({})\n\n{:?}", inbox, r.text().unwrap()),
             Err(e) => println!("Error while sending to inbox ({:?})", e)
         }
-    }
-}
-
-#[derive(Debug)]
-pub enum SignatureValidity {
-    Invalid,
-    ValidNoDigest,
-    Valid,
-    Absent,
-}
-
-impl SignatureValidity {
-    pub fn is_secure(&self) -> bool {
-        match self {
-            SignatureValidity::Valid => true,
-            _ => false,
-        }
-    }
-}
-
-pub fn verify_http_headers<S: sign::Signer+::std::fmt::Debug>(sender: &S, all_headers: HeaderMap, data: String) -> SignatureValidity{
-    if let Some(sig_header) = all_headers.get_one("Signature") {
-        let mut _key_id = None;
-        let mut _algorithm = None;
-        let mut headers = None;
-        let mut signature = None;
-        for part in sig_header.split(',') {
-            match part {
-                part if part.starts_with("keyId=") => _key_id = Some(&part[7..part.len()-1]),
-                part if part.starts_with("algorithm=") => _algorithm = Some(&part[11..part.len()-1]),
-                part if part.starts_with("headers=") => headers = Some(&part[9..part.len()-1]),
-                part if part.starts_with("signature=") => signature = Some(&part[11..part.len()-1]),
-                _ => {},
-            }
-        }
-        if signature.is_some() && headers.is_some() {
-            let headers = headers.unwrap().split_whitespace().collect::<Vec<_>>();
-            let signature = signature.unwrap();
-            let h = headers.iter()
-                .map(|header| (header,all_headers.get_one(header)))
-                .map(|(header, value)| format!("{}: {}", header.to_lowercase(), value.unwrap_or("")))
-                .collect::<Vec<_>>().join("\n");
-            if sender.verify(h, base64::decode(signature).unwrap_or(Vec::new())) {
-                if headers.contains(&"digest") {
-                    let digest = all_headers.get_one("digest").unwrap_or("");
-                    let digest = request::Digest::from_header(digest);
-                    if digest.map(|d| d.verify(data)).unwrap_or(false) {
-                        SignatureValidity::Valid
-                    } else {
-                        SignatureValidity::Invalid
-                    }
-                } else {
-                    SignatureValidity::ValidNoDigest
-                }
-            } else {
-                SignatureValidity::Invalid
-            }
-        } else {
-            SignatureValidity::Invalid
-        }
-    } else {
-        SignatureValidity::Absent
     }
 }
